Move opcodes2 to opcodes.go
